Add files via upload
Real Estate Visualization Dashboard



Real estate in Nigeria is a dynamic and rapidly growing sector, driven by factors such as population growth, urbanization, and economic development. Here are some key points about real estate in Nigeria:



1. Market Growth: Nigeria has one of the fastest-growing real estate markets in Africa. Urbanization is a major driver of this growth, with people moving from rural areas to cities in search of better opportunities.



2. Residential Sector: The demand for residential properties, including apartments, houses, and gated communities, is high, especially in major cities like Lagos, Abuja, and Port Harcourt. Developers are catering to different income brackets, from luxury to affordable housing.



3. Commercial Sector: Nigeria's growing economy is in high demand for commercial properties: retail outlets, office spaces, and industrial facilities. Major cities are witnessing the development of modern office complexes and shopping malls.



4. Infrastructure Development: Infrastructure plays a crucial role in the real estate sector. Investments in roads, bridges, power supply, and other amenities improve the attractiveness of properties and drive up their value.



5. Challenges: Despite its potential, the Nigerian real estate sector faces challenges such as inadequate infrastructure, lack of access to financing, cumbersome land acquisition processes, and issues related to property rights.



6. Regulation and Policy: The Nigerian government has been introducing policies and regulations aimed at stimulating growth in the real estate sector. 



7. Technology Adoption: The adoption of technology is transforming the real estate sector in Nigeria. Online property portals, digital payment platforms, and property management software are becoming increasingly popular, making it easier for buyers, sellers, and agents to transact.



Overall, real estate in Nigeria is a promising sector with significant growth potential, driven by demographic and economic factors. However, addressing infrastructure challenges and improving the regulatory environment are crucial for sustaining this growth and maximizing the sector's contribution to the economy.



The Real Estate data was obtained from #Kaggle, and it includes an analysis report on the real estate market in Nigeria. 





This dashboard displays information on different states, towns, and properties of built houses available for sale. It highlights the top locations in towns where houses are available in Nigeria. 



The cleaning and visualization dashboard is been analyze and done with #tableau published on #tableaupublic



The data not only provide information on the highest cost of those houses, but also offer potential locations for real estate agents to explore in Nigeria. 





The data is currently being updated and verified for authenticity. It is a work in progress and will be completed soon. 



#realestate #dataanalysis